+++ test/.eslintrc.js
@@ -1,4 +1,7 @@
 module.exports = {
+	plugins: [
+		'jest',
+	],
 	extends: [
 		'plugin:jest/recommended',
 		'plugin:jest/style',