+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 7.16-r1 / 2020-12-29
+* [**BREAKING**] JSDoc types are now expected to be lowercase.
+   * ✔ `@param {string} name The name.`
+   * ❌ `@param {String} name The name.`
+* Add module for **eslint-plugin-jsdoc**.
+* Add note in README about ESLint's ECMAScript Module support.
+
 ## 7.16-r0 / 2020-12-20
 * Update rules for ESLint 7.16.
 * Add module for **eslint-plugin-node**.